requirements.txt
databricks-sql-connector~=3.4.0
databricks-sdk~=0.33.0
streamlit~=1.38.0
pandas~=2.2.3
plotly~=5.24.0
PyYAML==6.0.2
pydantic==2.10.6